modules/location_producer/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
grpcio-tools
grpcio
